reorganize prompts to get rid of some errors
