+++ python/packages/ai/scripts/fmt.py
@@ -7,5 +7,5 @@ import subprocess
 
 
 def fmt():
-    subprocess.run(["poetry", "run", "black", "teams", "scripts", "tests", "examples"], check=True)
-    subprocess.run(["poetry", "run", "isort", "teams", "scripts", "tests", "examples"], check=True)
+    subprocess.run(["poetry", "run", "black", "teams", "scripts", "tests"], check=True)
+    subprocess.run(["poetry", "run", "isort", "teams", "scripts", "tests"], check=True)
