+++ python/packages/ai/scripts/lint.py
@@ -10,4 +10,3 @@ def lint():
     subprocess.run(["poetry", "run", "pylint", "teams", "scripts", "tests"], check=True)
     subprocess.run(["poetry", "run", "mypy", "--check-untyped-defs", "-p", "teams"], check=True)
     subprocess.run(["poetry", "run", "mypy", "--check-untyped-defs", "-p", "tests"], check=True)
-    subprocess.run(["poetry", "run", "mypy", "--check-untyped-defs", "-p", "examples"], check=True)
