+++ python/packages/ai/teams/ai/data_sources/data_source.py
@@ -7,7 +7,7 @@ from abc import ABC, abstractmethod
 
 from botbuilder.core import TurnContext
 
-from teams.ai.prompts.prompt_section import RenderedPromptSection
+from teams.ai.prompts.rendered_prompt_section import RenderedPromptSection
 from teams.ai.tokenizers import Tokenizer
 from teams.state.memory import Memory
