+++ python/packages/ai/teams/ai/prompts/function_call_message.py
@@ -14,8 +14,8 @@ from ..tokenizers import Tokenizer
 from .function_call import FunctionCall
 from .message import Message
 from .prompt_functions import PromptFunctions
-from .prompt_section_base import PromptSectionBase
 from .rendered_prompt_section import RenderedPromptSection
+from .sections.prompt_section_base import PromptSectionBase
 
 
 class FunctionCallMessage(PromptSectionBase):
