+++ python/packages/ai/tests/ai/prompts/test_layout_engine.py
@@ -8,14 +8,14 @@ from unittest import IsolatedAsyncioTestCase
 
 from botbuilder.core import TurnContext
 
-from teams.ai.prompts import LayoutEngine, PromptFunctions, TextSection
+from teams.ai.prompts import LayoutEngineSection, PromptFunctions, TextSection
 from teams.ai.tokenizers import GPTTokenizer
 from teams.state import TurnState
 
 
 class TestLayoutEngine(IsolatedAsyncioTestCase):
     async def test_render_as_text_empty_sections(self):
-        layout_engine = LayoutEngine([], 1, False, "\n")
+        layout_engine = LayoutEngineSection([], 1, False, "\n")
         result = await layout_engine.render_as_text(
             context=cast(TurnContext, {}),
             memory=TurnState(),
@@ -31,7 +31,7 @@ class TestLayoutEngine(IsolatedAsyncioTestCase):
     async def test_render_as_text_fixed_sections(self):
         section1 = TextSection("Hello World!", "user")
         section2 = TextSection("Teams-AI", "user")
-        layout_engine = LayoutEngine([section1, section2], 1, False, " ")
+        layout_engine = LayoutEngineSection([section1, section2], 1, False, " ")
         result = await layout_engine.render_as_text(
             context=cast(TurnContext, {}),
             memory=TurnState(),
@@ -47,7 +47,7 @@ class TestLayoutEngine(IsolatedAsyncioTestCase):
     async def test_render_as_text_fixed_sections_too_long(self):
         section1 = TextSection("Hello World!", "user")
         section2 = TextSection("Teams-AI", "user")
-        layout_engine = LayoutEngine([section1, section2], 1, False, " ")
+        layout_engine = LayoutEngineSection([section1, section2], 1, False, " ")
         result = await layout_engine.render_as_text(
             context=cast(TurnContext, {}),
             memory=TurnState(),
@@ -63,7 +63,7 @@ class TestLayoutEngine(IsolatedAsyncioTestCase):
     async def test_render_as_text_fixed_sections_drop_optional(self):
         section1 = TextSection("Hello World!", "user")
         section2 = TextSection("Teams-AI", "user", required=False)
-        layout_engine = LayoutEngine([section1, section2], 1, False, " ")
+        layout_engine = LayoutEngineSection([section1, section2], 1, False, " ")
         result = await layout_engine.render_as_text(
             context=cast(TurnContext, {}),
             memory=TurnState(),
@@ -79,7 +79,7 @@ class TestLayoutEngine(IsolatedAsyncioTestCase):
     async def test_render_as_text_fixed_sections_too_long_after_drop(self):
         section1 = TextSection("Hello World!", "user")
         section2 = TextSection("Teams-AI", "user", required=False)
-        layout_engine = LayoutEngine([section1, section2], 1, False, " ")
+        layout_engine = LayoutEngineSection([section1, section2], 1, False, " ")
         result = await layout_engine.render_as_text(
             context=cast(TurnContext, {}),
             memory=TurnState(),
@@ -95,7 +95,7 @@ class TestLayoutEngine(IsolatedAsyncioTestCase):
     async def test_render_as_text_proportional_sections(self):
         section1 = TextSection("Hello World!", "user", tokens=0.5)
         section2 = TextSection("Teams-AI", "user", tokens=0.5)
-        layout_engine = LayoutEngine([section1, section2], 1, False, " ")
+        layout_engine = LayoutEngineSection([section1, section2], 1, False, " ")
         result = await layout_engine.render_as_text(
             context=cast(TurnContext, {}),
             memory=TurnState(),
@@ -111,7 +111,7 @@ class TestLayoutEngine(IsolatedAsyncioTestCase):
     async def test_render_as_text_proportional_sections_drop_optional(self):
         section2 = TextSection("Teams-AI!", "user", tokens=0.5, required=False)
         section1 = TextSection("Hello World!", "user", tokens=0.5, required=False)
-        layout_engine = LayoutEngine([section1, section2], 1, False, " ")
+        layout_engine = LayoutEngineSection([section1, section2], 1, False, " ")
         result = await layout_engine.render_as_text(
             context=cast(TurnContext, {}),
             memory=TurnState(),
@@ -129,7 +129,7 @@ class TestLayoutEngine(IsolatedAsyncioTestCase):
     async def test_render_as_text_proportional_sections_too_long_after_drop(self):
         section1 = TextSection("Hello World!", "user", tokens=0.5)
         section2 = TextSection("Teams-AI!", "user", tokens=0.5, required=False)
-        layout_engine = LayoutEngine([section1, section2], 1, False, " ")
+        layout_engine = LayoutEngineSection([section1, section2], 1, False, " ")
         result = await layout_engine.render_as_text(
             context=cast(TurnContext, {}),
             memory=TurnState(),
@@ -143,7 +143,7 @@ class TestLayoutEngine(IsolatedAsyncioTestCase):
         self.assertTrue(result.too_long)
 
     async def test_render_as_messages_empty_sections(self):
-        layout_engine = LayoutEngine([], 1, False, "\n")
+        layout_engine = LayoutEngineSection([], 1, False, "\n")
         result = await layout_engine.render_as_messages(
             context=cast(TurnContext, {}),
             memory=TurnState(),
@@ -159,7 +159,7 @@ class TestLayoutEngine(IsolatedAsyncioTestCase):
     async def test_render_as_messages(self):
         section1 = TextSection("Hello World!", "user")
         section2 = TextSection("Teams-AI", "user")
-        layout_engine = LayoutEngine([section1, section2], 1, False, " ")
+        layout_engine = LayoutEngineSection([section1, section2], 1, False, " ")
         result = await layout_engine.render_as_messages(
             context=cast(TurnContext, {}),
             memory=TurnState(),
